+++ src/blob/file.rs
@@ -3,9 +3,15 @@ use std::{
     time::SystemTime,
 };
 
+use bytes::BytesMut;
 use nix::{errno::Errno, fcntl::FcntlArg};
 
 use super::prelude::*;
+
+// This size was selected based on results of tests
+// These tests were performed on 8-core machine with HDD
+// Results are confirmed on 8 and 80 threads, for put and get
+const MAX_SYNC_OPERATION_SIZE: usize = 100_000;
 
 #[derive(Debug, Clone)]
 pub struct File {
@@ -132,35 +138,56 @@ impl File {
     async fn write_at_sync(&self, offset: u64, buf: &[u8]) -> IOResult<()> {
         let buf = buf.to_vec();
         let file = self.no_lock_fd.clone();
-        Self::blocking_call(move || file.write_all_at(&buf, offset)).await
-    }
-
-    pub(crate) async fn read_all(&self) -> Result<Vec<u8>> {
-        let mut buf = vec![0; self.size().try_into()?];
-        self.read_at(&mut buf, 0).await?; // TODO: verify read size
-        Ok(buf)
-    }
-
-    pub(crate) async fn read_at(&self, buf: &mut [u8], offset: u64) -> Result<()> {
-        if let Some(ref ioring) = self.ioring {
-            self.read_at_aio(buf, offset, ioring).await
+        if buf.len() <= MAX_SYNC_OPERATION_SIZE {
+            Self::inplace_sync_call(move || file.write_all_at(&buf, offset))
         } else {
-            self.read_at_sync(buf, offset).await
+            Self::background_sync_call(move || file.write_all_at(&buf, offset)).await
         }
     }
 
-    pub(crate) async fn read_at_sync(&self, buf: &mut [u8], offset: u64) -> Result<()> {
-        let file = self.no_lock_fd.clone();
-        let mut new_buf = buf.to_vec();
-
-        let new_buf =
-            Self::blocking_call(move || file.read_exact_at(&mut new_buf, offset).map(|_| new_buf))
-                .await?;
-        buf.clone_from_slice(new_buf.as_slice());
-        Ok(())
+    pub(crate) async fn read_all(&self) -> Result<BytesMut> {
+        self.read_exact_at_allocate(self.size().try_into()?, 0)
+            .await
     }
 
-    async fn read_at_aio(&self, buf: &mut [u8], offset: u64, ioring: &Rio) -> Result<()> {
+    pub(crate) async fn read_exact_at_allocate(
+        &self,
+        size: usize,
+        offset: u64,
+    ) -> Result<BytesMut> {
+        let buf = BytesMut::zeroed(size);
+        self.read_exact_at(buf, offset).await
+    }
+
+    pub(crate) async fn read_exact_at(&self, buf: BytesMut, offset: u64) -> Result<BytesMut> {
+        if let Some(ref ioring) = self.ioring {
+            self.read_exact_at_aio(buf, offset, ioring).await
+        } else {
+            self.read_exact_at_sync(buf, offset).await
+        }
+    }
+
+    pub(crate) async fn read_exact_at_sync(
+        &self,
+        mut buf: BytesMut,
+        offset: u64,
+    ) -> Result<BytesMut> {
+        let file = self.no_lock_fd.clone();
+
+        Ok(if buf.len() <= MAX_SYNC_OPERATION_SIZE {
+            Self::inplace_sync_call(move || file.read_exact_at(&mut buf, offset).map(|_| buf))
+        } else {
+            Self::background_sync_call(move || file.read_exact_at(&mut buf, offset).map(|_| buf))
+                .await
+        }?)
+    }
+
+    async fn read_exact_at_aio(
+        &self,
+        mut buf: BytesMut,
+        offset: u64,
+        ioring: &Rio,
+    ) -> Result<BytesMut> {
         debug!("blob file read at");
         if buf.is_empty() {
             warn!("file read_at empty buf");
@@ -195,7 +222,7 @@ impl File {
             debug!("blob file read at proggress {}/{}", size, buf.len());
         }
         debug!("blob file read at complited: {}", size);
-        Ok(())
+        Ok(buf)
     }
 
     async fn from_tokio_file(file: TokioFile, ioring: Option<Rio>) -> IOResult<Self> {
@@ -217,7 +244,7 @@ impl File {
             compl.await
         } else {
             let fd = self.no_lock_fd.clone();
-            Self::blocking_call(move || fd.sync_all()).await
+            Self::background_sync_call(move || fd.sync_all()).await
         }
     }
 
@@ -241,7 +268,7 @@ impl File {
         }
     }
 
-    async fn blocking_call<F, R>(f: F) -> R
+    async fn background_sync_call<F, R>(f: F) -> R
     where
         F: FnOnce() -> R + Send + 'static,
         R: Send + 'static,
@@ -249,6 +276,14 @@ impl File {
         tokio::task::spawn_blocking(move || f())
             .await
             .expect("spawned blocking task failed")
+    }
+
+    fn inplace_sync_call<F, R>(f: F) -> R
+    where
+        F: FnOnce() -> R + Send + 'static,
+        R: Send + 'static,
+    {
+        tokio::task::block_in_place(move || f())
     }
 
     pub(crate) fn created_at(&self) -> Result<SystemTime> {