+++ CHANGELOG.md
@@ -7,7 +7,7 @@ Pearl changelog
 - Special handling for large data write (#192)
 
 #### Changed
-
+- Added different path for sync read/write (#184)
 
 #### Fixed